(1.25.1) Refactoring of c-level interfaces for RMA background, quantile normalization and median polish implementations
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/preprocessCore@83401 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ inst/include/preprocessCore_background_stubs.c
@@ -5,36 +5,36 @@
 #define PREPROCESSCORE_BACKGROUND_STUBS_H 1
 
 
-void rma_bg_parameters(double *PM,double *param, int rows, int cols, int column){
+void rma_bg_parameters(double *PM,double *param, size_t rows, size_t cols, size_t column){
 
-  static void(*fun)(double *, double *, int, int, int) = NULL;
+  static void(*fun)(double *, double *, size_t, size_t, size_t) = NULL;
 
   if (fun == NULL)
-    fun = (void(*)(double *, double *, int, int, int))R_GetCCallable("preprocessCore","rma_bg_parameters");
+    fun = (void(*)(double *, double *, size_t, size_t, size_t))R_GetCCallable("preprocessCore","rma_bg_parameters");
 
   fun(PM, param, rows, cols, column);
   return;
 }
 
 
-void rma_bg_adjust(double *PM,double *param, int rows, int cols, int column){
+void rma_bg_adjust(double *PM,double *param, size_t rows, size_t cols, size_t column){
 
-  static void(*fun)(double *, double *, int, int, int) = NULL;
+  static void(*fun)(double *, double *, size_t, size_t, size_t) = NULL;
 
   if (fun == NULL)
-    fun = (void(*)(double *, double *, int, int, int))R_GetCCallable("preprocessCore","rma_bg_adjust");
+    fun = (void(*)(double *, double *, size_t, size_t, size_t))R_GetCCallable("preprocessCore","rma_bg_adjust");
 
   fun(PM, param, rows, cols, column);
   return;
 }
 
 
-void rma_bg_correct(double *PM, int rows, int cols){
+void rma_bg_correct(double *PM, size_t rows, size_t cols){
 
-  static void(*fun)(double *, int, int) = NULL;
+  static void(*fun)(double *, size_t, size_t) = NULL;
 
   if (fun == NULL)
-    fun = (void(*)(double *, int, int))R_GetCCallable("preprocessCore","rma_bg_correct");
+    fun = (void(*)(double *, size_t, size_t))R_GetCCallable("preprocessCore","rma_bg_correct");
 
   fun(PM, rows, cols);
   return;
