+++ inst/include/preprocessCore_background_stubs.h
@@ -1,3 +1,3 @@
-void rma_bg_parameters(double *PM,double *param, int rows, int cols, int column);
-void rma_bg_adjust(double *PM,double *param, int rows, int cols, int column);
-void rma_bg_correct(double *PM, int rows, int cols);
+void rma_bg_parameters(double *PM,double *param, size_t rows, size_t cols, size_t column);
+void rma_bg_adjust(double *PM,double *param, size_t rows, size_t cols, size_t column);
+void rma_bg_correct(double *PM, size_t rows, size_t cols);
