+++ src/rma_background4.h
@@ -1,9 +1,9 @@
 #ifndef RMA_BACKGROUND4_H
 #define RMA_BACKGROUND4_H
 
-void rma_bg_parameters(double *PM,double *param, int rows, int cols, int column);
-void rma_bg_adjust(double *PM, double *param, int rows, int cols, int column);
-void rma_bg_correct(double *PM, int rows, int cols);
+void rma_bg_parameters(double *PM,double *param, size_t rows, size_t cols, size_t column);
+void rma_bg_adjust(double *PM, double *param, size_t rows, size_t cols, size_t column);
+void rma_bg_correct(double *PM, size_t rows, size_t cols);
 
 SEXP R_rma_bg_correct(SEXP PMmat,SEXP copy);
